Fix unit testing in Xcode
Unit testing appeared to be completely broken, so I removed and added again the xctest frameworks.
