+++ Project/Weight Tracker/Weight Tracker.xcodeproj/xcshareddata/xcschemes/Weight Tracker.xcscheme
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
    LastUpgradeVersion = "0700"
-   version = "1.3">
+   version = "1.7">
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
@@ -32,21 +32,29 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "FA49EC471B8F1FF10097AE8B"
-               BuildableName = "Weight TrackerTests.xctest"
-               BlueprintName = "Weight TrackerTests"
+               BlueprintIdentifier = "FADE26D81C012D7700A2AC09"
+               BuildableName = "Weight_TrackerTests.xctest"
+               BlueprintName = "Weight_TrackerTests"
                ReferencedContainer = "container:Weight Tracker.xcodeproj">
             </BuildableReference>
+            <LocationScenarioReference
+               identifier = "London, England"
+               referenceType = "1">
+            </LocationScenarioReference>
          </TestableReference>
          <TestableReference
-            skipped = "NO">
+            skipped = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "FA49EC521B8F1FF10097AE8B"
-               BuildableName = "Weight TrackerUITests.xctest"
-               BlueprintName = "Weight TrackerUITests"
+               BlueprintIdentifier = "FADA5D8C1C016AA1000CA9A2"
+               BuildableName = "Weight_TrackerUITests.xctest"
+               BlueprintName = "Weight_TrackerUITests"
                ReferencedContainer = "container:Weight Tracker.xcodeproj">
             </BuildableReference>
+            <LocationScenarioReference
+               identifier = "London, England"
+               referenceType = "1">
+            </LocationScenarioReference>
          </TestableReference>
       </Testables>
       <MacroExpansion>
