+++ Project/Weight Tracker/Weight Tracker.xcodeproj/xcuserdata/alex.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -27,6 +27,36 @@
 			<key>primary</key>
 			<true/>
 		</dict>
+		<key>FADA5D8C1C016AA1000CA9A2</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>FADE26BC1C012CF900A2AC09</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>FADE26CA1C012D0200A2AC09</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>FADE26D81C012D7700A2AC09</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>FADE26E61C012D7E00A2AC09</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>FADE26F41C012F9E00A2AC09</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
 	</dict>
 </dict>
 </plist>
